+++ rinex/src/tests/toolkit/observation.rs
@@ -152,7 +152,7 @@ pub fn generic_observation_rinex_test(
     assert_eq!(content, expected);
 
     // Self - Self should be 0
-    let null_dut = dut.substract(&dut);
+    let null_dut = dut.observation_substract(&dut);
     generic_null_rinex_test(&null_dut);
 
     // Check against provided model